# docs/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Tweek
  license:
    name: MIT
    url: 'https://github.com/Soluto/tweek/blob/master/LICENSE'
  version: 0.1.0
  description: |
    
    # Intro
    
    This is the openapi reference for Tweek api.
    Tweek provides api for reading configuration values, writing/reading context, managing keys/schemas/tags/policies etc...

    You can use openapi tools to generate different client to Tweek.

    # Security
    Tweek api support 2 models of authentication
    - Client Credentials - using headers x-client-id and x-client-secret, map to subject (external-apps {client-id})
    - JWT Token - Passed in Authorization header with Bearer scheme.  
        There are several ways to get JWT token
        - OIDC provider (Such as Google, IdentityServer, Auth0, etc...) -> map to subject based on subject extraction definition.
        - Signed JWT token with Repo private-key - issuer should be Tweek -> map to admin subject
        - JWT token create by basic auth using client Client Credentials -> map to subject (external-apps {client-id})
security:
  - bearerAuth: []
  - apiClientId: []
    apiClientSecret : []
paths:
...
  /apps:
    post:
      operationId: createApp
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppCreationResponseModel'
      summary: 'Create new app'
      description: |
        Create new app (credentials) for using Tweek.
        Return login details for the new app.

        require permission - write repo/apps 
      tags:
        - apps
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppCreationRequestModel'
        required: true
    get:
      operationId: getApps
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema: 
                type: object
      summary: 'List apps'
      description: |
        List all applications created in Tweek.  
        Return a map of appId->name.  

        require permission - read repo/apps 
      tags:
        - apps
...
servers:
  - url: /api/v2/
components:
  securitySchemes:
    bearerAuth:           
      type: http
      scheme: bearer
      bearerFormat: JWT   
    apiClientId:
      type: apiKey
      in: header
      name: X-CLIENT-ID
    apiClientSecret:
      type: apiKey
      in: header
      name: X-CLIENT-SECRET
...
  schemas:
...
    AppCreationResponseModel:
      description: ''
      properties:
        appId:
          type: string
          description: ''
        appSecret:
          type: string
          description: ''
      type: object
      required:
        - appId
        - appSecret
    AppCreationRequestModel:
      description: ''
      properties:
        name:
          type: string
          description: ''
        permissions:
          type: array
          items:
            type: string
          description: ''
      type: object
      required:
        - name
        - permissions